PenultimateDays.jl Documentation

PenultimateDays.jl Documentation

Adding PenultimateDays.jl

Documentation

Index

Adding PenultimateDays.jl

julia> using Pkg
julia> Pkg.add("PenultimateDays")    Updating registry at `~/.julia/registries/General.toml`
   Resolving package versions...
ERROR: expected package `PenultimateDays [556d87f4]` to be registered

Documentation

Trivial Penultimate Functions

Main.PenultimateDays.penultimatedayofweek — Method

penultimatedayofweek(dt::TimeType) -> TimeType

Adjusts dt to the penultimate (second-to-last) day of its week.

Main.PenultimateDays.penultimatedayofmonth — Method

penultimatedayofmonth(dt::TimeType) -> TimeType

Adjusts dt to the penultimate (second-to-last) day of its month.

Main.PenultimateDays.penultimatedayofquarter — Method

penultimatedayofquarter(dt::TimeType) -> TimeType

Adjusts dt to the penultimate (second-to-last) day of its quarter.

Main.PenultimateDays.penultimatedayofyear — Method

penultimatedayofyear(dt::TimeType) -> TimeType

Adjusts dt to the penultimate (second-to-last) day of its year.

Day-specific Penultimate Functions

Main.PenultimateDays.penultimatedayofmonth — Method

penultimatedayofmonth(dt::TimeType, d::Int) -> TimeType

Adjusts dt to the penultimate (second-to-last) day of its month, given some d, the day of the week as an Int, with 1 = Monday, 2 = Tuesday, &c...

For example, the penultimatedayofmonth(dt, 6) will find the penultimate Saturday of the month. Dates also exports integer aliases Monday–Sunday, so you can write penultimatedayofmonth(dt, Saturday).

See also: Dates.dayofweek, lastdayofmonth

Main.PenultimateDays.penultimatedayofquarter — Method

penultimatedayofquarter(dt::TimeType, d::Int) -> TimeType

Adjusts dt to the penultimate (second-to-last) day of its quarter, given some d, the day of the week as an Int, with 1 = Monday, 2 = Tuesday, &c...

For example, the penultimatedayofquarter(dt, 6) will find the penultimate Saturday of the quarter. Dates also exports integer aliases Monday–Sunday, so you can write penultimatedayofquarter(dt, Saturday).

See also: Dates.dayofweek, lastdayofquarter

Main.PenultimateDays.penultimatedayofyear — Method

penultimatedayofyear(dt::TimeType, d::Int) -> TimeType

Adjusts dt to the penultimate (second-to-last) day of its year, given some d, the day of the week as an Int, with 1 = Monday, 2 = Tuesday, &c...

For example, the penultimatedayofyear(dt, 6) will find the penultimate Saturday of the year. Dates also exports integer aliases Monday–Sunday, so you can write penultimatedayofyear(dt, Saturday).

See also: Dates.dayofweek, lastdayofyear

Dates (Standard Library) Extended

Missing docstring.

Missing docstring for firstdayofweek(::TimeType, ::Int). Check Documenter's build log for details.

Missing docstring.

Missing docstring for lastdayofweek(::TimeType, ::Int). Check Documenter's build log for details.

Missing docstring.

Missing docstring for firstdayofmonth(::TimeType, ::Int). Check Documenter's build log for details.

Missing docstring.

Missing docstring for lastdayofmonth(::TimeType, ::Int). Check Documenter's build log for details.

Missing docstring.

Missing docstring for firstdayofquarter(::TimeType, ::Int). Check Documenter's build log for details.

Missing docstring.

Missing docstring for lastdayofquarter(::TimeType, ::Int). Check Documenter's build log for details.

Missing docstring.

Missing docstring for firstdayofyear(::TimeType, ::Int). Check Documenter's build log for details.

Missing docstring.

Missing docstring for lastdayofyear(::TimeType, ::Int). Check Documenter's build log for details.

Index

Main.PenultimateDays.penultimatedayofmonth

Main.PenultimateDays.penultimatedayofmonth

Main.PenultimateDays.penultimatedayofquarter

Main.PenultimateDays.penultimatedayofquarter

Main.PenultimateDays.penultimatedayofweek

Main.PenultimateDays.penultimatedayofyear

Main.PenultimateDays.penultimatedayofyear