Documentation

Documentation

Trivial Penultimate Functions

Day-specific Penultimate Functions

Dates (Standard Library) Extended

Index

Examples

Basic examples

Extending the Dates standard library

PenultimateDays.jl Documentation

Adding PenultimateDays.jl

Trivial Penultimate Functions

Main.PenultimateDays.penultimatedayofweek — Method

penultimatedayofweek(dt::TimeType) -> TimeType

Adjusts dt to the penultimate (second-to-last) day of its week.

Main.PenultimateDays.penultimatedayofmonth — Method

penultimatedayofmonth(dt::TimeType) -> TimeType

Adjusts dt to the penultimate (second-to-last) day of its month.

Main.PenultimateDays.penultimatedayofquarter — Method

penultimatedayofquarter(dt::TimeType) -> TimeType

Adjusts dt to the penultimate (second-to-last) day of its quarter.

Main.PenultimateDays.penultimatedayofyear — Method

penultimatedayofyear(dt::TimeType) -> TimeType

Adjusts dt to the penultimate (second-to-last) day of its year.

Day-specific Penultimate Functions

Main.PenultimateDays.penultimatedayofmonth — Method

penultimatedayofmonth(dt::TimeType, d::Int) -> TimeType

Adjusts dt to the penultimate (second-to-last) day of its month, given some d, the day of the week as an Int, with 1 = Monday, 2 = Tuesday, &c...

For example, the penultimatedayofmonth(dt, 6) will find the penultimate Saturday of the month. Dates also exports integer aliases Monday–Sunday, so you can write penultimatedayofmonth(dt, Saturday).

See also: Dates.dayofweek, lastdayofmonth

Main.PenultimateDays.penultimatedayofquarter — Method

penultimatedayofquarter(dt::TimeType, d::Int) -> TimeType

Adjusts dt to the penultimate (second-to-last) day of its quarter, given some d, the day of the week as an Int, with 1 = Monday, 2 = Tuesday, &c...

For example, the penultimatedayofquarter(dt, 6) will find the penultimate Saturday of the quarter. Dates also exports integer aliases Monday–Sunday, so you can write penultimatedayofquarter(dt, Saturday).

See also: Dates.dayofweek, lastdayofquarter

Main.PenultimateDays.penultimatedayofyear — Method

penultimatedayofyear(dt::TimeType, d::Int) -> TimeType

Adjusts dt to the penultimate (second-to-last) day of its year, given some d, the day of the week as an Int, with 1 = Monday, 2 = Tuesday, &c...

For example, the penultimatedayofyear(dt, 6) will find the penultimate Saturday of the year. Dates also exports integer aliases Monday–Sunday, so you can write penultimatedayofyear(dt, Saturday).

See also: Dates.dayofweek, lastdayofyear

Dates (Standard Library) Extended

Dates.firstdayofweek — Function

firstdayofweek(dt::TimeType, d::Int) -> TimeType

Adjusts dt to the first d-day of its week, given some d, the day of the week as an Int, with 1 = Monday, 2 = Tuesday, &c...

For example, the firstdayofweek(dt, 6) will find the first Saturday of the week. Dates also exports integer aliases Monday–Sunday, so you can write firstdayofweek(dt, Saturday).

See also: Dates.dayofweek

Dates.lastdayofweek — Function

lastdayofweek(dt::TimeType, d::Int) -> TimeType

Adjusts dt to the last d-day of its week, given some d, the day of the week as an Int, with 1 = Monday, 2 = Tuesday, &c...

For example, the lastdayofweek(dt, 6) will find the last Saturday of the week. Dates also exports integer aliases Monday–Sunday, so you can write lastdayofweek(dt, Saturday).

See also: Dates.dayofweek

Dates.firstdayofmonth — Function

firstdayofmonth(dt::TimeType, d::Int) -> TimeType

Adjusts dt to the first d-day of its month, given some d, the day of the week as an Int, with 1 = Monday, 2 = Tuesday, &c...

For example, the firstdayofmonth(dt, 6) will find the first Saturday of the month. Dates also exports integer aliases Monday–Sunday, so you can write firstdayofmonth(dt, Saturday).

See also: Dates.dayofweek

Dates.lastdayofmonth — Function

lastdayofmonth(dt::TimeType, d::Int) -> TimeType

Adjusts dt to the last d-day of its month, given some d, the day of the week as an Int, with 1 = Monday, 2 = Tuesday, &c...

For example, the lastdayofmonth(dt, 6) will find the last Saturday of the month. Dates also exports integer aliases Monday–Sunday, so you can write lastdayofmonth(dt, Saturday).

See also: Dates.dayofweek

Dates.firstdayofquarter — Function

firstdayofquarter(dt::TimeType, d::Int) -> TimeType

Adjusts dt to the first d-day of its quarter, given some d, the day of the week as an Int, with 1 = Monday, 2 = Tuesday, &c...

For example, the firstdayofquarter(dt, 6) will find the first Saturday of the quarter. Dates also exports integer aliases Monday–Sunday, so you can write firstdayofquarter(dt, Saturday).

See also: Dates.dayofweek

Dates.lastdayofquarter — Function

lastdayofquarter(dt::TimeType, d::Int) -> TimeType

Adjusts dt to the last d-day of its quarter, given some d, the day of the week as an Int, with 1 = Monday, 2 = Tuesday, &c...

For example, the lastdayofquarter(dt, 6) will find the last Saturday of the quarter. Dates also exports integer aliases Monday–Sunday, so you can write lastdayofquarter(dt, Saturday).

See also: Dates.dayofweek

Dates.firstdayofyear — Function

firstdayofyear(dt::TimeType, d::Int) -> TimeType

Adjusts dt to the first d-day of its year, given some d, the day of the week as an Int, with 1 = Monday, 2 = Tuesday, &c...

For example, the firstdayofyear(dt, 6) will find the first Saturday of the year. Dates also exports integer aliases Monday–Sunday, so you can write firstdayofyear(dt, Saturday).

See also: Dates.dayofweek

Dates.lastdayofyear — Function

lastdayofyear(dt::TimeType, d::Int) -> TimeType

Adjusts dt to the last d-day of its year, given some d, the day of the week as an Int, with 1 = Monday, 2 = Tuesday, &c...

For example, the lastdayofyear(dt, 6) will find the last Saturday of the year. Dates also exports integer aliases Monday–Sunday, so you can write lastdayofyear(dt, Saturday).

See also: Dates.dayofweek

Index

Dates.firstdayofmonth

Dates.firstdayofquarter

Dates.firstdayofweek

Dates.firstdayofyear

Dates.lastdayofmonth

Dates.lastdayofquarter

Dates.lastdayofweek

Dates.lastdayofyear

Main.PenultimateDays.penultimatedayofmonth

Main.PenultimateDays.penultimatedayofmonth

Main.PenultimateDays.penultimatedayofquarter

Main.PenultimateDays.penultimatedayofquarter

Main.PenultimateDays.penultimatedayofweek

Main.PenultimateDays.penultimatedayofyear

Main.PenultimateDays.penultimatedayofyear